docs(852): move form-complite to separate story
